feat: Refactor project structure and integrate core features including map and ticket functionalities
