+++ frontend/lib/core/models/StopData.dart
