+++ frontend/lib/main.dart
@@ -1,401 +1,14 @@
+// lib/main.dart
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:flutter/services.dart' show rootBundle;
-import 'classes/StopData.dart';
-import 'dart:convert';
+import 'package:aurabus/app.dart';
+import 'package:aurabus/core/providers/app_state.dart';
 
 void main() {
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return ChangeNotifierProvider(
+  runApp(
+    ChangeNotifierProvider(
       create: (context) => MyAppState(),
-      child: MaterialApp(
-        title: 'AuraBus',
-        theme: ThemeData(
-          colorScheme: const ColorScheme(
-            brightness: Brightness.light,
-            primary: Colors.black,
-            onPrimary: Colors.black,
-            secondary: Colors.grey,
-            onSecondary: Colors.black,
-            error: Colors.red,
-            onError: Colors.black,
-            surface: Colors.grey,
-            onSurface: Colors.black,
-          ),
-        ),
-        home: const MyHomePage(),
-      ),
-    );
-  }
-}
-
-class MyAppState extends ChangeNotifier {}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key});
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  var selectedIndex = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    Widget page;
-    switch (selectedIndex) {
-      case 0:
-        page = const TicketPage();
-        break;
-      case 1:
-        page = const MapScreen();
-        break;
-      case 2:
-        page = const Placeholder();
-        break;
-      default:
-        throw UnimplementedError('no widget for $selectedIndex');
-    }
-
-    return Scaffold(
-      body: page,
-      bottomNavigationBar: BottomNavigationBar(
-        currentIndex: selectedIndex,
-        onTap: (value) {
-          setState(() {
-            selectedIndex = value;
-          });
-        },
-        items: const [
-          BottomNavigationBarItem(
-            icon: Icon(Icons.airplane_ticket),
-            label: 'Tickets',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.home),
-            label: 'Home',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.account_circle),
-            label: 'Account',
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-
-class MapScreen extends StatefulWidget {
-  const MapScreen({super.key});
-
-  @override
-  State<MapScreen> createState() => MapScreenState();
-}
-
-class MapScreenState extends State<MapScreen> {
-  late GoogleMapController mapController;
-  Set<Marker> _markers = {};
-  String? styleString;
-
-  final LatLng _center = const LatLng(46.067808715456785, 11.130308912105304);
-
-  void _onMapCreated(GoogleMapController controller) {
-    mapController = controller;
-    _loadMapStyle();
-    _loadMarkers();
-  }
-
-  void _loadMapStyle() async {
-    try {
-      final style = await rootBundle.loadString('assets/map_style.json');
-      setState(() {
-        styleString = style;
-      });
-      print("Stile mappa applicato con successo.");
-    } catch (e) {
-      print("Errore nel caricamento o applicazione dello stile: $e");
-    }
-  }
-
-  void _loadMarkers() async {
-    try {
-      final jsonString = await rootBundle.loadString('assets/stop_data.json');
-      final List<dynamic> stopsJson = json.decode(jsonString);
-      print("Elementi JSON trovati: ${stopsJson.length}");
-      final Set<Marker> newMarkers = {};
-
-      for (var stopJson in stopsJson) {
-        final stop = StopData.fromJson(stopJson);
-        final marker = Marker(
-          markerId: MarkerId(stop.stopId.toString()),
-          position: LatLng(stop.stopLat, stop.stopLon),
-          infoWindow: InfoWindow(
-            title: stop.stopName,
-            snippet:
-                'Linee: ${stop.routes.map((r) => r.routeShortName).join(', ')}',
-          ),
-        );
-        newMarkers.add(marker);
-      }
-
-      setState(() {
-        _markers = newMarkers;
-        print("Marker caricati: ${_markers.length}");
-      });
-    } catch (e) {
-      print("🚨 ERRORE nel caricamento/parsing dei marker: $e");
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: GoogleMap(
-        style: styleString,
-        onMapCreated: _onMapCreated,
-        initialCameraPosition: CameraPosition(target: _center, zoom: 13.0),
-        markers: _markers,
-        mapType: MapType.normal,
-        zoomControlsEnabled: true,
-      ),
-    );
-  }
-} 
-
-class TicketPage extends StatelessWidget {
-  const TicketPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xFFF8F8F8),
-      body: SafeArea(
-        child: SingleChildScrollView(
-          child: Column(
-            children: [
-              const SizedBox(height: 30),
-              const Text(
-                'Your Tickets',
-                style: TextStyle(fontSize: 26, fontWeight: FontWeight.bold),
-              ),
-              const SizedBox(height: 20),
-              const TicketCard(),
-              const SizedBox(height: 20),
-              const TicketCard(),
-              const SizedBox(height: 40),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class TicketCard extends StatelessWidget {
-  const TicketCard({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.circular(14),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.1),
-            blurRadius: 6,
-            offset: const Offset(0, 3),
-          ),
-        ],
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Image.asset(
-                'assets/tt_logo.png',
-                height: 32,
-                fit: BoxFit.contain,
-              ),
-              const Column(
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  Text(
-                    'Servizio Urbano',
-                    style: TextStyle(
-                      fontSize: 12,
-                      color: Colors.black54,
-                      fontWeight: FontWeight.w400,
-                    ),
-                  ),
-                  Text(
-                    'TRENTO',
-                    style: TextStyle(
-                      fontSize: 14,
-                      color: Colors.black87,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(width: 32),
-            ],
-          ),
-
-          const SizedBox(height: 12),
-
-          Row(
-            children: [
-              Expanded(
-                child: Container(height: 1, color: Colors.black12),
-              ),
-              const Padding(
-                padding: EdgeInsets.symmetric(horizontal: 6.0),
-                child: Text(
-                  'P.IVA 01807370224',
-                  style: TextStyle(fontSize: 10, color: Colors.black54),
-                ),
-              ),
-              Expanded(
-                child: Container(height: 1, color: Colors.black12),
-              ),
-            ],
-          ),
-
-          const SizedBox(height: 12),
-
-          Row(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              Expanded(
-                flex: 4,
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.stretch,
-                  children: [
-                    Row(
-                      children: [
-                        Expanded(
-                          child: Container(
-                            height: 42,
-                            decoration: const BoxDecoration(
-                              color: Color(0xFFE96E2B),
-                              borderRadius: BorderRadius.only(
-                                topLeft: Radius.circular(4),
-                                bottomLeft: Radius.circular(4),
-                              ),
-                            ),
-                            alignment: Alignment.center,
-                            child: const Text(
-                              '70 minuti',
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontWeight: FontWeight.bold,
-                                fontSize: 13,
-                              ),
-                            ),
-                          ),
-                        ),
-                        const SizedBox(width: 1.5),
-                        Expanded(
-                          child: Container(
-                            height: 42,
-                            decoration: const BoxDecoration(
-                              color: Color(0xFFE96E2B),
-                              borderRadius: BorderRadius.only(
-                                topRight: Radius.circular(4),
-                                bottomRight: Radius.circular(4),
-                              ),
-                            ),
-                            alignment: Alignment.center,
-                            child: const Text(
-                              '€ 1.20',
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontWeight: FontWeight.bold,
-                                fontStyle: FontStyle.italic,
-                                fontSize: 13,
-                              ),
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 12),
-                    Container(
-                      width: double.infinity,
-                      padding: const EdgeInsets.symmetric(vertical: 8),
-                      decoration: BoxDecoration(
-                        border: Border.all(color: Colors.black, width: 1.5),
-                        borderRadius: BorderRadius.circular(6),
-                      ),
-                      alignment: Alignment.center,
-                      child: const Text(
-                        'Validate',
-                        style: TextStyle(
-                          fontSize: 15,
-                          fontWeight: FontWeight.w600,
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              const SizedBox(width: 16),
-              Expanded(
-                flex: 2,
-                child: Column(
-                  children: [
-                    Container(
-                      padding: const EdgeInsets.symmetric(
-                        horizontal: 2,
-                        vertical: 2,
-                      ),
-                      decoration: BoxDecoration(
-                        color: Colors.grey[300],
-                        borderRadius: BorderRadius.circular(4),
-                      ),
-                      child: Column(
-                        children: [
-                          Image.asset(
-                            'assets/ticket_qr.png',
-                            width: 100,
-                            height: 100,
-                            fit: BoxFit.contain,
-                          ),
-                          const SizedBox(height: 4),
-                          const Text(
-                            'UNUSED',
-                            style: TextStyle(
-                              fontSize: 10,
-                              color: Colors.black87,
-                              letterSpacing: 0.5,
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
+      child: const MyApp(),
+    ),
+  );
 }
