+++ frontend/pubspec.yaml
@@ -1,4 +1,4 @@
-name: frontend
+name: aurabus
 description: "A new Flutter project."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
